Src upload

+++ customerservice/src/main/java/com/rabo/customerservice/dao/CustomerServiceDAO.java
@@ -10,15 +10,10 @@ import com.rabo.customerservice.model.CustomerModel;
 public interface CustomerServiceDAO {
 
 	public Optional<CustomerModel> addCustomer(CustomerModel customer);
-
-	Optional<CustomerModel> getCustomer(int id);
-
+	Optional<CustomerModel> getCustomer(Long id);
 	List<CustomerModel> getAllCustomers();
-
 	List<CustomerModel> getCustomerByFnameAndLname(String fname, String lname);
-
 	List<CustomerModel> getCustomerByFnameOrLname(String fname, String lname);
-
-	public CustomerModel updateLivingAddress(int custId, AddressModel address);
+	public CustomerModel updateLivingAddress(Long custId, AddressModel address);
 
 }
